Move all encrypted related classes from properties.h to new files: encryption_properties.h and encryption_properties.cc
