+++ cpp/src/parquet/properties.h
@@ -21,10 +21,7 @@
 #include <memory>
 #include <string>
 #include <unordered_map>
-#include <random>
-#include <openssl/rand.h>
 
-#include "parquet/encryption.h"
 #include "parquet/exception.h"
 #include "parquet/parquet_version.h"
 #include "parquet/platform.h"
@@ -32,6 +29,8 @@
 #include "parquet/types.h"
 #include "arrow/util/logging.h"
 #include "arrow/util/utf8.h"
+#include "parquet/encryption_properties.h"
+
 namespace parquet {
 
 struct ParquetVersion {
@@ -40,369 +39,6 @@ struct ParquetVersion {
 
 static int64_t DEFAULT_BUFFER_SIZE = 0;
 static bool DEFAULT_USE_BUFFERED_STREAM = false;
-static constexpr bool DEFAULT_CHECK_SIGNATURE = true;
-static const std::string NULL_STRING = "";
-
-class PARQUET_EXPORT ColumnEncryptionProperties {
- public:
-  class Builder {
-   public:
-    // Convenience builder for regular (not nested) columns.
-    Builder(const std::string name) {
-      Builder(schema::ColumnPath::FromDotString(name), true);
-    }
-
-    // Convenience builder for encrypted columns.
-    Builder(const std::shared_ptr<schema::ColumnPath>& path)
-      : Builder(path, true) {}
-
-    // Set a column-specific key.
-    // If key is not set on an encrypted column, the column will
-    // be encrypted with the footer key.
-    // keyBytes Key length must be either 16, 24 or 32 bytes.
-    Builder* withKey(const std::string& key) {
-      if (key.empty ())
-        return this;
-
-      DCHECK(!key.empty());
-      key_ = key;
-      return this;
-    }
-
-    // Set a key retrieval metadata.
-    // use either withKeyMetaData or withKeyID, not both
-    Builder* withKeyMetaData(const std::string& key_metadata) {
-      DCHECK(!key_metadata.empty());
-      DCHECK(key_metadata_.empty());
-      key_metadata_ = key_metadata;
-      return this;
-    }
-
-    // Set a key retrieval metadata (converted from String).
-    // use either withKeyMetaData or withKeyID, not both
-    // key_id will be converted to metadata (UTF-8 array).
-    Builder* withKeyID(std::string key_id) {
-      //key_id is expected to be in UTF8 encoding
-      ::arrow::util::InitializeUTF8();
-      const uint8_t *data = reinterpret_cast<const uint8_t*>(key_id.c_str());
-      if (!::arrow::util::ValidateUTF8(data, key_id.size()))
-        throw ParquetException("key id should be in UTF8 encoding");
-
-      DCHECK(!key_id.empty());
-      this->withKeyMetaData(key_id);
-      return this;
-    }
-
-    std::shared_ptr<ColumnEncryptionProperties> build() {
-      return std::shared_ptr<ColumnEncryptionProperties>(
-          new ColumnEncryptionProperties(encrypted_,
-                                         column_path_,
-                                         key_,
-                                         key_metadata_));
-    }
-
-   private:
-    const std::shared_ptr<schema::ColumnPath> column_path_;
-    bool encrypted_;
-    std::string key_;
-    std::string key_metadata_;
-
-    Builder(const std::shared_ptr<schema::ColumnPath>& path, bool encrypted)
-      : column_path_(path), encrypted_(encrypted) {}
-  };
-  const std::shared_ptr<schema::ColumnPath>& getPath() { return column_path_; }
-  bool isEncrypted() const { return encrypted_; }
-  bool isEncryptedWithFooterKey() const { return encrypted_with_footer_key_; }
-  const std::string& getKey() const { return key_; }
-  const std::string& getKeyMetaData() const { return key_metadata_; }
-
-  ColumnEncryptionProperties() = default;
-  ColumnEncryptionProperties(const ColumnEncryptionProperties& other) = default;
-  ColumnEncryptionProperties(ColumnEncryptionProperties&& other) = default;
-
- private:
-  const std::shared_ptr<schema::ColumnPath> column_path_;
-  bool encrypted_;
-  bool encrypted_with_footer_key_;
-  std::string key_;
-  std::string key_metadata_;
-  explicit ColumnEncryptionProperties(
-      bool encrypted,
-      const std::shared_ptr<schema::ColumnPath>& column_path,
-      const std::string& key,
-      const std::string& key_metadata):column_path_(column_path){
-    DCHECK(column_path != nullptr);
-    if (!encrypted)
-      DCHECK(key.empty() && key_metadata.empty());
-
-    if (!key.empty())
-      DCHECK(key.length() == 16 || key.length() == 24 || key.length() == 32);
-
-    encrypted_with_footer_key_ = (encrypted && key.empty());
-    if (encrypted_with_footer_key_)
-      DCHECK(key_metadata.empty());
-
-    encrypted_ = encrypted;
-    key_metadata_ = key_metadata;
-    key_ = key;
-  }
-};
-
-class PARQUET_EXPORT ColumnDecryptionProperties {
- public:
-  class Builder {
-   public:
-    // convenience builder for regular (not nested) columns.
-    Builder(const std::string name) {
-      Builder(schema::ColumnPath::FromDotString(name));
-    }
-
-    Builder(const std::shared_ptr<schema::ColumnPath>& path)
-      : column_path_(path) {}
-
-    // Set an explicit column key. If applied on a file that contains
-    // key metadata for this column the metadata will be ignored,
-    // the column will be decrypted with this key.
-    // key length must be either 16, 24 or 32 bytes.
-    Builder* withKey(const std::string& key) {
-      if (key.empty ())
-        return this;
-
-      DCHECK(!key.empty());
-      key_ = key;
-      return this;
-    }
-
-    std::shared_ptr<ColumnDecryptionProperties> build() {
-      return std::shared_ptr<ColumnDecryptionProperties>(
-          new ColumnDecryptionProperties(column_path_, key_));
-    }
-
-  private:
-    const std::shared_ptr<schema::ColumnPath> column_path_;
-    std::string key_;
-  };
-
-  ColumnDecryptionProperties() = default;
-  ColumnDecryptionProperties(const ColumnDecryptionProperties& other) = default;
-  ColumnDecryptionProperties(ColumnDecryptionProperties&& other) = default;
-
-  const std::shared_ptr<schema::ColumnPath>& getPath() { return column_path_; }
-  const std::string& getKey() const { return key_; }
-
- private:
-  const std::shared_ptr<schema::ColumnPath> column_path_;
-  std::string key_;
-
-  // This class is only required for setting explicit column decryption keys -
-  // to override key retriever (or to provide keys when key metadata and/or
-  // key retriever are not available)
-  explicit ColumnDecryptionProperties(
-      const std::shared_ptr<schema::ColumnPath>& column_path,
-      const std::string& key):column_path_(column_path){
-    DCHECK(column_path != nullptr);
-
-    if (!key.empty())
-      DCHECK(key.length() == 16 || key.length() == 24 || key.length() == 32);
-
-    key_ = key;
-  }
-};
-
-class PARQUET_EXPORT AADPrefixVerifier {
- public:
-  // Verifies identity (AAD Prefix) of individual file,
-  // or of file collection in a data set.
-  // Throws exception if an AAD prefix is wrong.
-  // In a data set, AAD Prefixes should be collected,
-  // and then checked for missing files.
-  virtual void check(std::string aad_prefix) = 0;
-};
-
-class PARQUET_EXPORT FileDecryptionProperties {
- public:
-  class Builder {
-  public:
-    Builder(){
-      check_plaintext_footer_integrity_ = DEFAULT_CHECK_SIGNATURE;
-    }
-
-    // Set an explicit footer key. If applied on a file that contains
-    // footer key metadata the metadata will be ignored, the footer
-    // will be decrypted/verified with this key.
-    // If explicit key is not set, footer key will be fetched from
-    // key retriever.
-    // param footerKey Key length must be either 16, 24 or 32 bytes.
-    Builder* withFooterKey(std::string footer_key) {
-      if (footer_key.empty ()) {
-        return this;
-      }
-      DCHECK(!footer_key.empty());
-      footer_key_ = footer_key;
-      return this;
-    }
-
-    // Set explicit column keys (decryption properties).
-    // Its also possible to set a key retriever on this property object.
-    // Upon file decryption, availability of explicit keys is checked before
-    // invocation of the retriever callback.
-    // If an explicit key is available for a footer or a column,
-    // its key metadata will be ignored.
-    Builder* withColumnKeys(const std::map<std::shared_ptr<schema::ColumnPath>,
-                            std::shared_ptr<ColumnDecryptionProperties>,
-                            schema::ColumnPath::CmpColumnPath>&
-                            column_properties) {
-      if (column_properties.size () == 0)
-        return this;
-
-      if (column_property_map_.size () != 0)
-        throw ParquetException("Column properties already set");
-
-      column_property_map_ = column_properties;
-      return this;
-    }
-
-    // Set a key retriever callback. Its also possible to
-    // set explicit footer or column keys on this file property object.
-    // Upon file decryption, availability of explicit keys is checked before
-    // invocation of the retriever callback.
-    // If an explicit key is available for a footer or a column,
-    // its key metadata will be ignored.
-    Builder* withKeyRetriever(const std::shared_ptr<DecryptionKeyRetriever>&
-                              key_retriever) {
-      if (key_retriever == NULLPTR)
-        return this;
-
-      DCHECK(key_retriever_ == NULLPTR);
-      key_retriever_ = key_retriever;
-      return this;
-    }
-
-    // Skip integrity verification of plaintext footers.
-    // If not called, integrity of plaintext footers will be checked in runtime,
-    // and an exception will be thrown in the following situations:
-    // - footer signing key is not available
-    // (not passed, or not found by key retriever)
-    // - footer content and signature don't match
-    Builder* withoutFooterSignatureVerification() {
-      check_plaintext_footer_integrity_ = false;
-      return this;
-    }
-
-    // Explicitly supply the file AAD prefix.
-    // A must when a prefix is used for file encryption, but not stored in file.
-    // If AAD prefix is stored in file, it will be compared to the explicitly
-    // supplied value and an exception will be thrown if they differ.
-    Builder* withAADPrefix(std::string aad_prefix) {
-      if (aad_prefix.empty()) {
-        return this;
-      }
-      DCHECK(aad_prefix_.empty());
-      aad_prefix_ = aad_prefix;
-      return this;
-    }
-
-    // Set callback for verification of AAD Prefixes stored in file.
-    Builder* withAADPrefixVerifier(
-        std::shared_ptr<AADPrefixVerifier> aad_prefix_verifier) {
-      if (aad_prefix_verifier == NULLPTR)
-        return this;
-
-      DCHECK(aad_prefix_verifier_ == NULLPTR);
-      aad_prefix_verifier_ = aad_prefix_verifier;
-      return this;
-    }
-
-    std::shared_ptr<FileDecryptionProperties> build() {
-      return std::shared_ptr<FileDecryptionProperties>(
-          new FileDecryptionProperties(footer_key_,
-                                       key_retriever_,
-                                       check_plaintext_footer_integrity_,
-                                       aad_prefix_,
-                                       aad_prefix_verifier_,
-                                       column_property_map_));
-    }
-
-  private:
-    std::string footer_key_;
-    std::string aad_prefix_;
-    std::shared_ptr<AADPrefixVerifier> aad_prefix_verifier_;
-
-    std::map<std::shared_ptr<schema::ColumnPath>,
-      std::shared_ptr<ColumnDecryptionProperties>,
-      schema::ColumnPath::CmpColumnPath> column_property_map_;
-
-    std::shared_ptr<DecryptionKeyRetriever> key_retriever_;
-    bool check_plaintext_footer_integrity_;
-  };
-
-  const std::string& getColumnKey(
-      const std::shared_ptr<schema::ColumnPath>& column_path) {
-    if (column_property_map_.find(column_path) != column_property_map_.end()) {
-      auto column_prop = column_property_map_[column_path];
-      if (column_prop != nullptr)
-        return column_prop->getKey();
-    }
-    return NULL_STRING;
-  }
-
-  const std::string& getFooterKey() {
-    return footer_key_;
-  }
-
-  const std::string& getAADPrefix() { return aad_prefix_; }
-
-  std::shared_ptr<DecryptionKeyRetriever> getKeyRetriever() {
-    return key_retriever_;
-  }
-
-  bool checkFooterIntegrity() {
-    return check_plaintext_footer_integrity_;
-  }
-
-  const std::shared_ptr<AADPrefixVerifier> &getAADPrefixVerifier() {
-    return aad_prefix_verifier_;
-  }
-
- private:
-  std::string footer_key_;
-  std::string aad_prefix_;
-  std::shared_ptr<AADPrefixVerifier> aad_prefix_verifier_;
-
-  std::map<std::shared_ptr<schema::ColumnPath>,
-    std::shared_ptr<ColumnDecryptionProperties>,
-    schema::ColumnPath::CmpColumnPath> column_property_map_;
-
-  std::shared_ptr<DecryptionKeyRetriever> key_retriever_;
-  bool check_plaintext_footer_integrity_;
-
-
-  FileDecryptionProperties(
-      const std::string& footer_key,
-      const std::shared_ptr<DecryptionKeyRetriever>& key_retriever,
-      bool check_plaintext_footer_integrity,
-      std::string aad_prefix,
-      std::shared_ptr<AADPrefixVerifier> aad_prefix_verifier,
-      std::map<std::shared_ptr<schema::ColumnPath>,
-      std::shared_ptr<ColumnDecryptionProperties>,
-      schema::ColumnPath::CmpColumnPath> column_property_map) {
-    DCHECK(!footer_key.empty() ||
-           NULLPTR != key_retriever ||
-           0 != column_property_map.size());
-    if (!footer_key.empty())
-      DCHECK(footer_key.length() == 16 || footer_key.length() == 24 ||
-             footer_key.length() == 32);
-    if (footer_key.empty() && check_plaintext_footer_integrity)
-      DCHECK(NULLPTR != key_retriever);
-    aad_prefix_verifier_ = aad_prefix_verifier;
-    footer_key_ = footer_key;
-    check_plaintext_footer_integrity_ = check_plaintext_footer_integrity;
-    key_retriever_ = key_retriever;
-    aad_prefix_ = aad_prefix;
-    column_property_map_ = column_property_map;
-  }
-
-};
 
 class PARQUET_EXPORT ReaderProperties {
  public:
@@ -471,10 +107,6 @@ static constexpr ParquetVersion::type DEFAULT_WRITER_VERSION =
     ParquetVersion::PARQUET_1_0;
 static const char DEFAULT_CREATED_BY[] = CREATED_BY_VERSION;
 static constexpr Compression::type DEFAULT_COMPRESSION_TYPE = Compression::UNCOMPRESSED;
-static constexpr ParquetCipher::type DEFAULT_ENCRYPTION_ALGORITHM = ParquetCipher::AES_GCM_V1;
-static constexpr int32_t MAXIMAL_AAD_METADATA_LENGTH = 256;
-static constexpr bool DEFAULT_ENCRYPTED_FOOTER = true;
-static constexpr int32_t AAD_FILE_UNIQUE_LENGTH = 8;
 
 class PARQUET_EXPORT ColumnProperties {
  public:
@@ -521,208 +153,6 @@ class PARQUET_EXPORT ColumnProperties {
   bool dictionary_enabled_;
   bool statistics_enabled_;
   size_t max_stats_size_;
-};
-
-class PARQUET_EXPORT FileEncryptionProperties {
- public:
-  class Builder {
-  public:
-   Builder(const std::string& footer_key)
-     : parquet_cipher_(DEFAULT_ENCRYPTION_ALGORITHM),
-      encrypted_footer_(DEFAULT_ENCRYPTED_FOOTER) {
-      footer_key_ = footer_key;
-      store_aad_prefix_in_file_ = false;
-    }
-
-    // Create files with plaintext footer.
-    // If not called, the files will be created with encrypted footer (default).
-    Builder* withPlaintextFooter() {
-      encrypted_footer_ = false;
-      return this;
-    }
-
-    // Set encryption algorithm.
-    // If not called, files will be encrypted with AES_GCM_V1 (default).
-    Builder* withAlgorithm(ParquetCipher::type parquet_cipher) {
-      parquet_cipher_ = parquet_cipher;
-      return this;
-    }
-
-    // Set a key retrieval metadata (converted from String).
-    // use either withFooterKeyMetaData or withFooterKeyID, not both.
-    Builder* withFooterKeyID(std::string key_id) {
-      //key_id is expected to be in UTF8 encoding
-      ::arrow::util::InitializeUTF8();
-      const uint8_t* data = reinterpret_cast<const uint8_t*>(key_id.c_str());
-      if (!::arrow::util::ValidateUTF8(data, key_id.size()))
-        throw ParquetException("footer key id should be in UTF8 encoding");
-
-      if (key_id.empty())
-        return this;
-
-      return withFooterKeyMetadata(key_id);
-    }
-
-    // Set a key retrieval metadata.
-    // use either withFooterKeyMetaData or withFooterKeyID, not both.
-    Builder* withFooterKeyMetadata(const std::string& footer_key_metadata) {
-      if (footer_key_metadata.empty())
-        return this;
-
-      DCHECK(footer_key_metadata_.empty());
-      footer_key_metadata_ = footer_key_metadata;
-      return this;
-    }
-
-    // Set the file AAD Prefix.
-    Builder* withAADPrefix(const std::string& aad_prefix) {
-      if (aad_prefix.empty())
-        return this;
-
-      DCHECK(aad_prefix_.empty());
-      aad_prefix_ = aad_prefix;
-      store_aad_prefix_in_file_ = true;
-      return this;
-    }
-
-    // Skip storing AAD Prefix in file.
-    // If not called, and if AAD Prefix is set, it will be stored.
-    Builder* withoutAADPrefixStorage() {
-      store_aad_prefix_in_file_ = false;
-      return this;
-    }
-
-    // Set the list of encrypted columns and their properties (keys etc).
-    // If not called, all columns will be encrypted with the footer key.
-    // If called, the file columns not in the list will be left unencrypted.
-    Builder* withEncryptedColumns(
-        const std::map<std::shared_ptr<schema::ColumnPath>,
-        std::shared_ptr<ColumnEncryptionProperties>,
-        schema::ColumnPath::CmpColumnPath>&
-        encryptedColumns){
-      if (encryptedColumns.size () == 0)
-        return this;
-
-      if (column_property_map_.size () != 0)
-        throw ParquetException("Column properties already set");
-
-      column_property_map_ = encryptedColumns;
-      return this;
-    }
-
-    std::shared_ptr<FileEncryptionProperties> build() {
-      return std::shared_ptr<FileEncryptionProperties>(
-          new FileEncryptionProperties(parquet_cipher_,
-                                       footer_key_,
-                                       footer_key_metadata_,
-                                       encrypted_footer_,
-                                       aad_prefix_,
-                                       store_aad_prefix_in_file_,
-                                       column_property_map_));
-    }
-
-  private:
-    ParquetCipher::type parquet_cipher_;
-    bool encrypted_footer_;
-    std::string footer_key_;
-    std::string footer_key_metadata_;
-
-    std::string aad_prefix_;
-    bool store_aad_prefix_in_file_;
-    std::map<std::shared_ptr<schema::ColumnPath>,
-      std::shared_ptr<ColumnEncryptionProperties>,
-      schema::ColumnPath::CmpColumnPath> column_property_map_;
-  };
-
-  bool encryptedFooter() const { return encrypted_footer_; }
-
-  const EncryptionAlgorithm getAlgorithm() {
-    return algorithm_;
-  }
-
-  const std::string& getFooterEncryptionKey() {
-    return (encrypted_footer_? footer_key_ : NULL_STRING);
-  }
-
-  const std::string& getFooterEncryptionKeyMetadata() {
-    return (encrypted_footer_? footer_key_metadata_ : NULL_STRING);
-  }
-
-  const std::string& getFooterSigningKey() {
-    return (encrypted_footer_? NULL_STRING : footer_key_);
-  }
-
-  const std::string& getFooterSigningKeyMetadata() {
-    return (encrypted_footer_? NULL_STRING : footer_key_metadata_);
-  }
-
-  const std::string& getFileAAD() const { return file_AAD_; }
-
-  std::shared_ptr<ColumnEncryptionProperties> getColumnProperties(
-      const std::shared_ptr<schema::ColumnPath>& column_path) {
-    if (column_property_map_.size () == 0){
-      auto builder = std::shared_ptr<ColumnEncryptionProperties::Builder>(
-          new ColumnEncryptionProperties::Builder (column_path));
-      return builder->build();
-    }
-    if (column_property_map_.find(column_path) != column_property_map_.end())
-      return column_property_map_[column_path];
-
-    return NULLPTR;
-  }
-
- private:
-  EncryptionAlgorithm algorithm_;  // encryption algorithm
-  std::string footer_key_;
-  std::string footer_key_metadata_;
-  bool encrypted_footer_;
-  std::string file_AAD_;
-
-  std::map<std::shared_ptr<schema::ColumnPath>,
-    std::shared_ptr<ColumnEncryptionProperties>,
-    schema::ColumnPath::CmpColumnPath> column_property_map_;
-
-  FileEncryptionProperties(ParquetCipher::type cipher,
-                           std::string footer_key,
-                           std::string footer_key_metadata,
-                           bool encrypted_footer,
-                           const std::string& aad_prefix,
-                           bool store_aad_prefix_in_file,
-                           const std::map<std::shared_ptr<schema::ColumnPath>,
-                           std::shared_ptr<ColumnEncryptionProperties>,
-                           schema::ColumnPath::CmpColumnPath>&
-                           column_property_map)
-    : footer_key_(footer_key),
-    footer_key_metadata_(footer_key_metadata),
-    encrypted_footer_(encrypted_footer),
-    column_property_map_(column_property_map){
-      DCHECK(!footer_key.empty());
-      // footer_key must be either 16, 24 or 32 bytes.
-      DCHECK(footer_key.length() == 16
-             || footer_key.length() == 24
-             || footer_key.length() == 32);
-
-      uint8_t aad_file_unique[AAD_FILE_UNIQUE_LENGTH];
-      memset(aad_file_unique, 0, AAD_FILE_UNIQUE_LENGTH);
-      RAND_bytes(aad_file_unique, sizeof(AAD_FILE_UNIQUE_LENGTH));
-      std::string aad_file_unique_str(
-          reinterpret_cast<char const*>(aad_file_unique),
-          AAD_FILE_UNIQUE_LENGTH) ;
-
-      bool supply_aad_prefix = false;
-      if (aad_prefix.empty())
-        file_AAD_ = aad_file_unique_str;
-      else {
-        file_AAD_ = aad_prefix + aad_file_unique_str;
-        if (!store_aad_prefix_in_file) supply_aad_prefix = true;
-      }
-      algorithm_.algorithm = cipher;
-      algorithm_.aad.aad_file_unique = aad_file_unique_str;
-      algorithm_.aad.supply_aad_prefix = supply_aad_prefix;
-      if (!aad_prefix.empty() && store_aad_prefix_in_file) {
-        algorithm_.aad.aad_prefix = aad_prefix;
-      }
-    }
 };
 
 class PARQUET_EXPORT WriterProperties {
